# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"CountryCode": {
				"allOf": [
					{
						"type": "string"
					},
					{
						"properties": {
							"__length3__": {
								"type": "boolean",
								"enum": [
									true
								],
								"nullable": false
							}
						},
						"required": [
							"__length3__"
						],
						"type": "object"
					}
				],
				"description": "A branded type for country codes.\n\nThis type ensures that strings representing country codes have exactly 3 characters."
			},
...
			"WorldRegion": {
				"description": "Enum representing the different world regions.",
				"enum": [
					"Africa",
					"Americas",
					"Asia",
					"Europe",
					"Oceania",
					"Antarctica",
					"Polar"
				],
				"type": "string"
			},
			"WorldSubregion": {
				"description": "Enum representing the various subregions as defined in the restcountries API.",
				"enum": [
					"Northern Africa",
					"Sub-Saharan Africa",
					"Latin America and the Caribbean",
					"Caribbean",
					"Central America",
					"South America",
					"Northern America",
					"Central Asia",
					"Eastern Asia",
					"South-Eastern Asia",
					"Southern Asia",
					"Western Asia",
					"Eastern Europe",
					"Northern Europe",
					"Southern Europe",
					"Western Europe",
					"Australia and New Zealand",
					"Melanesia",
					"Micronesia",
					"Polynesia",
					"Antarctica"
				],
				"type": "string"
			},
			"Url": {
				"allOf": [
					{
						"type": "string"
					},
					{
						"properties": {
							"__brand": {
								"type": "string",
								"enum": [
									"url"
								],
								"nullable": false
							}
						},
						"required": [
							"__brand"
						],
						"type": "object"
					}
				]
			},
			"Candidate": {
				"description": "Represents a candidate country for hosting the FIFA World Cup.\nCombines key data from both the team and country perspectives.",
				"properties": {
					"countryCode": {
						"$ref": "#/components/schemas/CountryCode",
						"description": "The ISO 3166-1 alpha-3 country code."
					},
					"countryName": {
						"type": "string",
						"description": "The official name of the country."
					},
					"area": {
						"$ref": "#/components/schemas/PositiveNumber",
						"description": "The total area of the country in square kilometers."
					},
					"population": {
						"$ref": "#/components/schemas/PositiveNumber",
						"description": "The population of the country."
					},
					"region": {
						"$ref": "#/components/schemas/WorldRegion",
						"description": "The region (continent) where the country is located (e.g., Europe, Americas)."
					},
					"subregion": {
						"$ref": "#/components/schemas/WorldSubregion",
						"description": "The subregion (e.g., Caribbean, Melanesia) the country is part of."
					},
					"capital": {
						"type": "string",
						"description": "The capital city of the country."
					},
					"teamName": {
						"type": "string",
						"description": "The name of the team, usually the name of the country it represents."
					},
					"teamIconSvgUrl": {
						"$ref": "#/components/schemas/Url",
						"description": "The URL to the team's icon or flag image in SVG format."
					},
					"countryFlagSvgUrl": {
						"$ref": "#/components/schemas/Url",
						"description": "The URL to the team's icon or flag image in SVG format."
					}
				},
				"required": [
					"countryCode",
					"countryName",
					"area",
					"population",
					"region",
					"subregion",
					"capital",
					"teamName",
					"teamIconSvgUrl",
					"countryFlagSvgUrl"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CountryHeader": {
				"description": "Represents a country's header information including general geographic and demographic data.",
				"properties": {
					"countryName": {
						"type": "string",
						"description": "The name of the country."
					},
					"countryCode": {
						"$ref": "#/components/schemas/CountryCode",
						"description": "The ISO 3166-1 alpha-3 country code."
					},
					"area": {
						"$ref": "#/components/schemas/PositiveNumber",
						"description": "The total area of the country in square kilometers."
					},
					"population": {
						"$ref": "#/components/schemas/PositiveNumber",
						"description": "The population of the country."
					},
					"region": {
						"$ref": "#/components/schemas/WorldRegion",
						"description": "The region (continent) the country is part of (e.g., Europe, Asia)."
					},
					"subregion": {
						"$ref": "#/components/schemas/WorldSubregion",
						"description": "The subregion (e.g., Caribbean, Melanesia) the country is part of."
					},
					"countryFlagSvgUrl": {
						"$ref": "#/components/schemas/Url",
						"description": "The URL to the team's icon or flag image in SVG format."
					}
				},
				"required": [
					"countryName",
					"countryCode",
					"area",
					"population",
					"region",
					"subregion",
					"countryFlagSvgUrl"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TeamHeader": {
				"description": "Represents the essential header information for a FIFA team.",
				"properties": {
					"countryCode": {
						"$ref": "#/components/schemas/CountryCode",
						"description": "The ISO 3166-1 alpha-3 country code, used for cross-referencing with country data."
					},
					"teamName": {
						"type": "string",
						"description": "The name of the team, usually the name of the country it represents."
					},
					"teamIconSvgUrl": {
						"$ref": "#/components/schemas/Url",
						"description": "The URL to the team's icon or flag image in SVG format."
					}
				},
				"required": [
					"countryCode",
					"teamName",
					"teamIconSvgUrl"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "fifa-wwc",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Bojan Misic ",
			"email": "misic.bojan@gmail.com"
		}
	},
	"paths": {
		"/candidates": {
			"get": {
				"operationId": "GetAllCandidates",
				"responses": {
					"200": {
						"description": "An array of `Candidate` objects.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Candidate"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves all candidates based on the provided query parameters.",
				"summary": "Get the list of all FIFA World Cup 2024 candidates.",
				"tags": [
					"Candidates"
				],
				"security": [],
				"parameters": [
					{
						"description": "- The minimum area to filter candidates by (positive number) in square kilometers.",
						"in": "query",
						"name": "minArea",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- The maximum area to filter candidates by (positive number) in square kilometers.",
						"in": "query",
						"name": "maxArea",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- The minimum population to filter candidates by (positive number) of population.",
						"in": "query",
						"name": "minPopulation",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- The maximum population to filter candidates by (positive number) of population.",
						"in": "query",
						"name": "maxPopulation",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- The region (continent) to filter candidates by.",
						"in": "query",
						"name": "region",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/WorldRegion"
						}
					},
					{
						"description": "- The subregion (e.g., Caribbean, Melanesia) to filter candidates by.",
						"in": "query",
						"name": "subregion",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/WorldSubregion"
						}
					},
					{
						"description": "- The property to sort candidates by.",
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"countryCode",
								"countryName",
								"area",
								"population",
								"region",
								"subregion",
								"capital",
								"teamName",
								"teamIconSvgUrl",
								"countryFlagSvgUrl"
							]
						}
					},
					{
						"description": "- The order to sort candidates by (asc/desc).",
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					}
				]
			}
		},
		"/candidates/country": {
			"get": {
				"operationId": "GetAllCountries",
				"responses": {
					"200": {
						"description": "An array of `CountryHeader` objects.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CountryHeader"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves all countries on the provided query parameters.",
				"summary": "Get the list of all FIFA World Cup 2024 countries with essential information.",
				"tags": [
					"Countries"
				],
				"security": [],
				"parameters": [
					{
						"description": "- The region (continent) to filter countries by.",
						"in": "query",
						"name": "region",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/WorldRegion"
						}
					},
					{
						"description": "- The subregion (e.g., Caribbean, Melanesia) to filter countries by.",
						"in": "query",
						"name": "subregion",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/WorldSubregion"
						}
					},
					{
						"description": "- The property to sort countries by.",
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"countryCode",
								"countryName",
								"area",
								"population",
								"region",
								"subregion",
								"countryFlagSvgUrl"
							]
						}
					},
					{
						"description": "- The order to sort countries by (asc/desc).",
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					}
				]
			}
		},
		"/candidates/country/details": {
			"get": {
				"operationId": "GetAllCountriesDetails",
				"responses": {
					"200": {
						"description": "An array of `CountryDetails` objects.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CountryDetails"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves all countries details based on the provided query parameters.",
				"summary": "Get the list of all FIFA World Cup 2024 countries with detailed information.",
				"tags": [
					"Countries"
				],
				"security": [],
				"parameters": [
					{
						"description": "- The region (continent) to filter countries by.",
						"in": "query",
						"name": "region",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/WorldRegion"
						}
					},
					{
						"description": "- The subregion (e.g., Caribbean, Melanesia) to filter countries by.",
						"in": "query",
						"name": "subregion",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/WorldSubregion"
						}
					},
					{
						"description": "- The minimum population to filter countries by (positive number) of population.",
						"in": "query",
						"name": "minPopulation",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- The maximum population to filter countries by (positive number) of population.",
						"in": "query",
						"name": "maxPopulation",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- The minimum area to filter countries by (positive number) in square kilometers.",
						"in": "query",
						"name": "minArea",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- The maximum area to filter countries by (positive number) in square kilometers.",
						"in": "query",
						"name": "maxArea",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- The minimum longitude to filter countries by (positive number) in degrees.",
						"in": "query",
						"name": "minLongitude",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- The maximum longitude to filter countries by (positive number) in degrees.",
						"in": "query",
						"name": "maxLongitude",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- The minimum latitude to filter countries by (positive number) in degrees.",
						"in": "query",
						"name": "minLatitude",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- The maximum latitude to filter countries by (positive number) in degrees.",
						"in": "query",
						"name": "maxLatitude",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "- The timezones to filter countries by (example: \"UTC-01:00\", \"UTC\", \"UTC+01:00\", \"UTC+02:00\", etc.).",
						"in": "query",
						"name": "timezones",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"description": "- The borders (country codes) to filter countries by (example: \"FR\", \"ES\", \"IT\", etc.).",
						"in": "query",
						"name": "borders",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"description": "- The languages to filter countries by (example: \"French\", \"Italian\", \"English\", etc.).",
						"in": "query",
						"name": "languages",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"description": "- The currencies to filter countries by (example: \"EUR\", \"USD\", \"GBP\", etc.).",
						"in": "query",
						"name": "currencies",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"description": "- The property to sort countries by.",
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"countryCode",
								"countryName",
								"area",
								"population",
								"region",
								"subregion",
								"capital",
								"languages",
								"currencies",
								"latitude",
								"longitude",
								"timezones",
								"borders",
								"independent"
							]
						}
					},
					{
						"description": "- The order to sort countries by (asc/desc).",
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					}
				]
			}
		},
		"/candidates/team": {
			"get": {
				"operationId": "GetAllTeams",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TeamHeader"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Teams"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"countryCode",
								"teamName",
								"teamIconSvgUrl"
							]
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					}
				]
			}
		},
		"/candidates/team/details": {
			"get": {
				"operationId": "GetAllTeamDetails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TeamDetails"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Teams"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "minGroupRank",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxGroupRank",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minTotalRank",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxTotalRank",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minGoalsScored",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxGoalsScored",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minGoalsConceded",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "maxGoalsConceded",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"countryCode",
								"teamName",
								"teamIconSvgUrl",
								"groupRank",
								"totalRank",
								"goalsScored",
								"goalsConceded"
							]
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					}
				]
			}
		},
		"/candidates/{countryCode}": {
			"get": {
				"operationId": "GetCandidateByCountryCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Candidate"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Candidates"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "countryCode",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/candidates/{countryCode}/country": {
			"get": {
				"operationId": "GetCountryHeaderByCountryCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/CountryHeader"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Countries"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "countryCode",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/candidates/{countryCode}/country/details": {
			"get": {
				"operationId": "GetCountryDetailsByCountryCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/CountryDetails"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Countries"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "countryCode",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/candidates/{countryCode}/team": {
			"get": {
				"operationId": "GetTeamHeaderByCountryCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/TeamHeader"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Teams"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "countryCode",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/candidates/{countryCode}/team/details": {
			"get": {
				"operationId": "GetTeamDetailsByCountryCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/TeamDetails"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Teams"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "countryCode",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}